# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MentorMinds Stellar API",
    "version": "1.0.0",
    "description": "Backend API for the MentorMinds platform — connecting mentors and mentees with Stellar blockchain payments.",
    "contact": {
      "name": "MentorMinds Team",
      "email": "support@mentorminds.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.mentorminds.com/api/v1",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /auth/login"
      }
    },
...
  },
  "paths": {
...
    "/admin/payments": {
      "get": {
        "summary": "List all payments with optional date range filter",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of payments"
          }
        }
      }
    },
...
    "/admin/analytics/payments": {
      "get": {
        "summary": "Payment volume and method breakdown",
        "tags": [
          "Admin Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "7d",
                "30d",
                "90d",
                "1y"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment volume and method breakdown"
          }
        }
      }
    },
...
    "/payments/webhook": {
      "post": {
        "summary": "Handle Stellar webhook events",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "transaction_hash": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed"
          }
        }
      }
    },
    "/payments": {
      "post": {
        "summary": "Initiate a payment",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Unique key to prevent duplicate payments"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bookingId",
                  "amount"
                ],
                "properties": {
                  "bookingId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "amount": {
                    "type": "string",
                    "example": "100.0000000"
                  },
                  "currency": {
                    "type": "string",
                    "default": "XLM"
                  },
                  "description": {
                    "type": "string"
                  },
                  "fromAddress": {
                    "type": "string"
                  },
                  "toAddress": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment initiated"
          }
        }
      },
      "get": {
        "summary": "List user payments",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "completed",
                "failed",
                "cancelled",
                "refunded"
              ]
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "payment",
                "refund",
                "platform_fee",
                "mentor_payout",
                "escrow_hold",
                "escrow_release"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payments"
          }
        }
      }
    },
    "/payments/history": {
      "get": {
        "summary": "Get payment history (completed payments only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment history with total volume"
          }
        }
      }
    },
    "/payments/{id}": {
      "get": {
        "summary": "Get payment details",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment details"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/payments/{id}/status": {
      "get": {
        "summary": "Check payment status",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status"
          }
        }
      }
    },
    "/payments/{id}/confirm": {
      "post": {
        "summary": "Confirm payment with Stellar transaction hash",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "stellarTxHash"
                ],
                "properties": {
                  "stellarTxHash": {
                    "type": "string",
                    "pattern": "^[a-fA-F0-9]{64}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment confirmed"
          }
        }
      }
    },
    "/payments/{id}/refund": {
      "post": {
        "summary": "Process a refund",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  },
                  "stellarTxHash": {
                    "type": "string",
                    "pattern": "^[a-fA-F0-9]{64}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund processed"
          }
        }
      }
    },
...
  },
...
}